Gateway Refactory
I really want to put most of the "business logic" into one place with
te gateway. Also, I was doing some GoLang no-no's by having multiple
writers to a shared map with the registry and gateway. This puts everything
back into the gateway and makes the registry just a specialized cache.

There is still more work to do with figuring out how to organize logic
between the registry and the gateway (maybe they just belong in the same
set of stuffs??), but things are looking much better.